:white_check_mark: [Test] 모든 친구 id 조회 internal API 컨트롤러 테스트
